requirements.txt
opencv-contrib-python
numpy
pillow
matplotlib